improve examples
