--- example/watcher/index.html
@@ -1,9 +1,0 @@
-<html>
-  <body>
-    <script src='/event.js'></script>
-    <script src='/wsf.js'></script>
-    <script type="text/javascript">
-      
-    </script>
-  </body>
-</html>
