+++ examples/file-watcher/index.html
@@ -1,0 +1,9 @@
+<html>
+  <body>
+    <script src='/event.js'></script>
+    <script src='/wsf.js'></script>
+    <script type="text/javascript">
+      
+    </script>
+  </body>
+</html>
