+++ examples/file-watcher/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "watcher",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "start": "node ./bin/www"
+  },
+  "dependencies": {
+    "express": "~4.9.0",
+    "body-parser": "~1.8.1",
+    "cookie-parser": "~1.3.3",
+    "morgan": "~1.3.0",
+    "serve-favicon": "~2.1.3",
+    "debug": "~2.0.0",
+    "ejs": "~0.8.5",
+    "fslider_ws": "*"
+  }
+}
