+++ examples/file-watcher/views/error.ejs
@@ -1,0 +1,3 @@
+<h1><%= message %></h1>
+<h2><%= error.status %></h2>
+<pre><%= error.stack %></pre>
