+++ examples/misc/ws3.html
@@ -1,0 +1,20 @@
+<html>
+<body>
+  <img id="ws">
+  <script src="/event.js"></script>
+  <script src="/wsf.js"></script>
+  <script>
+    /* reference usage */
+    var url = location.href.split('://')[1];
+    ws=wsf.connect('ws://' + url, function (socket) {
+      // set reconnect no-delay
+      socket.autoreconnect = false;
+      // normal close
+      //socket.close(1000);
+      socket.on('close', function (info) {
+        console.log(info);
+      });
+    });
+  </script>
+</body>
+</html>
