--- test/npm-debug.log
@@ -1,24 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ 'node', '/usr/local/bin/npm', 'install' ]
-2 info using npm@2.7.0
-3 info using node@v1.5.0
-4 verbose node symlink /usr/local/bin/node
-5 error install Couldn't read dependencies
-6 verbose stack Error: Invalid version: "0.5"
-6 verbose stack     at Object.module.exports.fixVersionField (/usr/local/lib/node_modules/npm/node_modules/normalize-package-data/lib/fixer.js:190:13)
-6 verbose stack     at /usr/local/lib/node_modules/npm/node_modules/normalize-package-data/lib/normalize.js:30:38
-6 verbose stack     at Array.forEach (native)
-6 verbose stack     at normalize (/usr/local/lib/node_modules/npm/node_modules/normalize-package-data/lib/normalize.js:29:15)
-6 verbose stack     at final (/usr/local/lib/node_modules/npm/node_modules/read-package-json/read-json.js:367:33)
-6 verbose stack     at then (/usr/local/lib/node_modules/npm/node_modules/read-package-json/read-json.js:126:33)
-6 verbose stack     at /usr/local/lib/node_modules/npm/node_modules/read-package-json/read-json.js:331:40
-6 verbose stack     at FSReqWrap.readFileAfterClose [as oncomplete] (evalmachine.<anonymous>:379:3)
-7 verbose cwd /home/ran/github/RocketEngine/test
-8 error Linux 3.13.0-46-generic
-9 error argv "node" "/usr/local/bin/npm" "install"
-10 error node v1.5.0
-11 error npm  v2.7.0
-12 error Invalid version: "0.5"
-13 error If you need help, you may report this error at:
-13 error     <http://github.com/npm/npm/issues>
-14 verbose exit [ 1, true ]
